移植版最终版v1.0

+++ USER/24C02/24c02.h
@@ -23,13 +23,13 @@ void IIC_write_OneByte(u8 Senddata);
 u8 IIC_Read_OneByte(u8 ack);
 	
 //EEPROM24c02相关函数
-u8 AT24C02_ReadByte(u8 ReadAddr);							     //指定地址读取一个字节
-void AT24C02_WriteByte(u8 WriteAddr,u8 DataToWrite);		//指定地址写入一个字节
+u8 AT24C02_ReadByte(u16 ReadAddr);							     //指定地址读取一个字节
+void AT24C02_WriteByte(u16 WriteAddr,u8 DataToWrite);		//指定地址写入一个字节
 
 u32 Buf_4Byte(u8 *pBuffer,u32 Date_4Byte,u8 Byte_num,u8 mode);
 
-void AT24C02_Write(u8 WriteAddr,u8 *pBuffer,u8 WriteNum);	//从指定地址开始写入指定长度的数据
-void AT24C02_Read(u8 ReadAddr,u8 *pBuffer,u8 ReadNum);   	//从指定地址开始读出指定长度的数据
+void AT24C02_Write(u16 WriteAddr,u8 *pBuffer,u16 WriteNum);	//从指定地址开始写入指定长度的数据
+void AT24C02_Read(u16 ReadAddr,u8 *pBuffer,u16 ReadNum);   	//从指定地址开始读出指定长度的数据
 
 u8 AT24C02_Test(void);  //检查器件
 void AT24C02_Init(void); //初始化IIC
